food-bot — копия (2)/requirements.txt
aiogram>=3.13
python-dotenv>=1.0.0
aiosqlite>=0.20